remove redundant implementation of normalized_array
